refactor: 로그아웃 함수 이동
